backend/requirements.txt
Flask
flask-cors
torch
transformers
openai-whisper
symspellpy
sounddevice
numpy
scipy
datasets
sentencepiece
accelerate
protobuf
pillow
mss
opencv-python
pytesseract
pyautogui
python-dotenv
requests
pymongo
dnspython
bcrypt
